1) naming refactoring 2) add props to a table and it's components 3) HeaderCell should be within a header page because it's uses only there
