+++ forza-horizon-community.front/src/components/CarTableComponent/components/SortingTableHead/constants.ts
@@ -1,5 +1,5 @@
 import { ICar } from "@/data-transfer-objects";
-import { IHeaderCell } from "../types";
+import { IHeaderCell } from "../../types";
 
 export const headerCells: IHeaderCell<ICar>[] = [
   {
