+++ forza-horizon-community.front/src/components/CarTableComponent/components/SortingTableHead/index.ts
@@ -1,3 +1,4 @@
 export { default as SortingTableHead } from "./SortingTableHead";
 
+export { headerCells } from "./constants";
 export * from "./types";
