+++ forza-horizon-community.front/src/components/CarTableComponent/components/SortingTableHead/types.ts
@@ -1,9 +1,11 @@
-import { IHeaderCell } from "@/components";
+import { TableHeadProps } from "@mui/material";
 
 export type OrderDirection = "asc" | "desc";
 
-export interface ISortingTableHeaderProps<TEntity> extends ISortingTableHeaderHook<TEntity> {
-  headerCells: IHeaderCell<TEntity>[];
+export interface IHeaderCell<TEntity> {
+  id: keyof TEntity;
+  lable: string;
+  isSortable: boolean;
 }
 
 export interface ISortingTableHeaderHook<TEntity> {
@@ -12,4 +14,9 @@ export interface ISortingTableHeaderHook<TEntity> {
   setOrder: (order: OrderDirection) => void;
   setOrderBy: (property: keyof TEntity) => void;
   sortEntities: (newOrder: OrderDirection, newProperty: keyof TEntity) => void;
+}
+
+export interface ISortingTableHeaderProps<TEntity> extends ISortingTableHeaderHook<TEntity> {
+  headerCells: IHeaderCell<TEntity>[];
+  props?: TableHeadProps;
 }
