+++ forza-horizon-community.front/src/components/CarTableComponent/components/index.ts
@@ -1,3 +1,1 @@
 export * from "./SortingTableHead";
-
-export { headerCells } from "./constants";
