+++ forza-horizon-community.front/src/components/CarTableComponent/index.ts
@@ -1,4 +1,3 @@
 export { default as CarTableComponent } from "./CarTableComponent";
 
 export * from "./components";
-export * from "./types";
