--- forza-horizon-community.front/src/components/CarTableComponent/types.ts
@@ -1,5 +1,0 @@
-export interface IHeaderCell<TEntity> {
-  id: keyof TEntity;
-  lable: string;
-  isSortable: boolean;
-}
