+++ forza-horizon-community.front/src/components/constants/types.ts
@@ -1,4 +1,4 @@
-export interface IDefaultCarThumbnailsize {
+export interface IImageSize {
   width: number;
   height: number;
 }
